# Speed up event store ingestion + benchmark