--- a/crates/sui-storage/Cargo.toml
+++ b/crates/sui-storage/Cargo.toml
@@ -36,6 +36,7 @@ eyre = "0.6.8"
 
 [dev-dependencies]
 anyhow = "1.0.64"
+criterion = { version = "0.3.6", features = ["async", "async_tokio"] }
 tempfile = "3.3.0"
 num_cpus = "1.13.1"
 pretty_assertions = "1.2.0"
@@ -43,4 +44,8 @@ telemetry-subscribers.workspace = true
 
 [[bench]]
 name = "write_ahead_log"
+harness = false
+
+[[bench]]
+name = "event_store_bench"
 harness = false