--- a/crates/sui-storage/src/event_store/sql.rs
+++ b/crates/sui-storage/src/event_store/sql.rs
@@ -16,7 +16,7 @@ use sui_types::base_types::SuiAddress;
 use sui_types::object::Owner;
 
 use sqlx::{
-    sqlite::{SqliteConnectOptions, SqliteRow},
+    sqlite::{SqliteConnectOptions, SqliteJournalMode, SqliteRow, SqliteSynchronous},
     Executor, Row, SqlitePool,
 };
 use sui_types::error::SuiError;
@@ -108,6 +108,10 @@ impl SqlEventStore {
         // TODO: configure other SQLite options
         let mut options = SqliteConnectOptions::new()
             .filename(db_path)
+            // SQLite turns off WAL by default and uses DELETE journaling.  WAL is at least 2x faster.
+            .journal_mode(SqliteJournalMode::Wal)
+            // Normal vs Full sync mode also speeds up writes
+            .synchronous(SqliteSynchronous::Normal)
             .create_if_missing(true);
         options.log_statements(log::LevelFilter::Off);
         let pool = SqlitePool::connect_with(options)