We found that event store writes might be very slow in production full nodes.  It was discovered default SQLite configuration turns off WAL and has a per-commit sync policy, which is too slow esp for EBS type stores.  This adds a benchmark and benchmarks various improvements.

1. Add a benchmark (see output below) using Criterion for event store ingestion
2. Turns on WAL journal mode and normal synchronization.  This is shown using the benchmark to speed up writes by 120% or so.
3. Switches to using batch inserts instead of individual inserts.  For number of events > 1, has a significant boost to write speeds.

Originally, for all batch sizes (1 event to 100 events), write speeds locally benchmarked at ~2000 events/sec (1 event) to 1800 (100 events).  Now:

(TL/DR; optimum event batch size is around 50, when we get >8500 events/sec, more than 4x the original mark)

```
Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /1: Collecting 100 samples i                                                                                                                         SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /1
                        time:   [19.783 ms 20.148 ms 20.579 ms]
                        thrpt:  [4.8592 Kelem/s 4.9633 Kelem/s 5.0549 Kelem/s]
                 change:
                        time:   [-5.4342% -2.9743% -0.5123%] (p = 0.02 < 0.05)
                        thrpt:  [+0.5149% +3.0655% +5.7465%]
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe
Event store event count: 65500
Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /5: Collecting 100 samples i                                                                                                                         SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /5
                        time:   [13.586 ms 13.841 ms 14.111 ms]
                        thrpt:  [7.0869 Kelem/s 7.2248 Kelem/s 7.3607 Kelem/s]
                 change:
                        time:   [-37.589% -35.635% -33.771%] (p = 0.00 < 0.05)
                        thrpt:  [+50.992% +55.364% +60.229%]
                        Performance has improved.
Event store event count: 131000
Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /10: Collecting 100 samples                                                                                                                          SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /10
                        time:   [12.706 ms 12.936 ms 13.178 ms]
                        thrpt:  [7.5884 Kelem/s 7.7305 Kelem/s 7.8700 Kelem/s]
                 change:
                        time:   [-43.952% -42.145% -40.505%] (p = 0.00 < 0.05)
                        thrpt:  [+68.081% +72.847% +78.420%]
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  4 (4.00%) low severe
  5 (5.00%) high mild
  9 (9.00%) high severe
Event store event count: 196500
Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /20: Collecting 100 samples                                                                                                                          SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /20
                        time:   [12.075 ms 12.377 ms 12.711 ms]
                        thrpt:  [7.8670 Kelem/s 8.0795 Kelem/s 8.2817 Kelem/s]
                 change:
                        time:   [-47.061% -45.238% -43.521%] (p = 0.00 < 0.05)
                        thrpt:  [+77.056% +82.610% +88.896%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
Event store event count: 272000
Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /50: Collecting 100 samples                                                                                                                          SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /50
                        time:   [11.475 ms 11.726 ms 11.999 ms]
                        thrpt:  [8.3341 Kelem/s 8.5278 Kelem/s 8.7146 Kelem/s]
                 change:
                        time:   [-48.718% -47.287% -45.826%] (p = 0.00 < 0.05)
                        thrpt:  [+84.591% +89.707% +95.000%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
Event store event count: 347500
Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /100: Warming up for 3.0000                                                                                                                          Benchmarking SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /100: Collecting 100 samples                                                                                                                         SQLite ingestion varying batch size/SqlEventStore.add_events() with batch size: /100
                        time:   [11.687 ms 12.044 ms 12.419 ms]
                        thrpt:  [8.0522 Kelem/s 8.3031 Kelem/s 8.5567 Kelem/s]
                 change:
                        time:   [-48.997% -47.119% -45.238%] (p = 0.00 < 0.05)
                        thrpt:  [+82.608% +89.105% +96.068%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```

Note that this PR does not batch events from multiple TXes, so this only benefits writes when transactions contain multiple events.  However, this can be turned on subsequently for another nice boost.